simplify swaggerInitializer
