+++ lib/routes.js
@@ -4,7 +4,7 @@ const path = require('path')
 const yaml = require('yaml')
 const fastifyStatic = require('@fastify/static')
 const rfdc = require('rfdc')()
-const serialize = require('./serialize')
+const swaggerInitializer = require('./swaggerInitializer')
 
 // URI prefix to separate static assets for swagger UI
 const staticPrefix = '/static'
@@ -75,45 +75,7 @@ function fastifySwagger (fastify, opts, done) {
     }
   })
 
-  const uiConfig = serialize(opts.uiConfig)
-  const initOAuth = serialize(opts.initOAuth)
-
-  const swaggerInitializer = `window.onload = function () {
-  function resolveUrl(url) {
-    const anchor = document.createElement('a')
-    anchor.href = url
-    return anchor.href
-  }
-
-  function resolveConfig(cb) {
-    const config = ${uiConfig}
-    const resConfig = Object.assign({}, {
-      dom_id: '#swagger-ui',
-      deepLinking: true,
-      presets: [
-        SwaggerUIBundle.presets.apis,
-        SwaggerUIStandalonePreset
-      ],
-      plugins: [
-        SwaggerUIBundle.plugins.DownloadUrl
-      ],
-      layout: "StandaloneLayout"
-    }, config, {
-      url: resolveUrl('./json').replace('static/json', 'json'),
-      oauth2RedirectUrl: resolveUrl('./oauth2-redirect.html')
-    });
-    return cb(resConfig);
-  }
-
-  // Begin Swagger UI call region
-  const buildUi = function (config) {
-    const ui = SwaggerUIBundle(config)
-    window.ui = ui
-    ui.initOAuth(${initOAuth});
-  }
-  // End Swagger UI call region
-  resolveConfig(buildUi);
-}`
+  const swaggerInitializerContent = swaggerInitializer(opts)
 
   fastify.route({
     url: '/static/swagger-initializer.js',
@@ -123,7 +85,7 @@ function fastifySwagger (fastify, opts, done) {
     handler: (req, reply) => {
       reply
         .header('content-type', 'application/javascript; charset=utf-8')
-        .send(swaggerInitializer)
+        .send(swaggerInitializerContent)
     }
   })
